requirements.txt
numpy==1.18.1
#opencv-python==4.2.0.34
opencv-python==4.3.0.38
networkx==2.4
matplotlib==3.3.0
scipy==1.5.2
tqdm==4.48.2
tifffile==2020.9.3
pytransform3d==1.9.1
Shapely==1.7.1
pyserial
jupyterlab

# for testing
deepdiff==5.7.0
hypothesis[numpy]==6.31.6

# additional for Mask R-CNN
scikit-image==0.16.2
cython==0.29.21
#nvidia-pyindex
#nvidia-tensorflow[hovorod]
keras==2.1.5
h5py==2.10.0
imgaug==0.2.9
IPython[all]

# additional for OpenSfM
cloudpickle==0.4.0
exifread==2.1.2
flask==1.1.2
fpdf2==2.4.2
joblib==0.14.1
Pillow==8.3.1
pyproj==2.5.0
pytest==3.0.7
python-dateutil==2.8.2
pyyaml==5.4
Sphinx==3.4.3
six==1.16.0
xmltodict==0.10.2
wheel
